LayoutTests/fast/css/style-scoped/style-scoped-in-shadow-expected.txt

Tests that

PASS window.internals is defined.
PASS window.internals.ensureShadowRoot is defined.
PASS WebKitShadowRoot is defined.
PASS shadow.innerHTML is defined.
PASS shadow.getElementById is defined.
A: rgb(0, 0, 0)
B: rgb(0, 0, 0)
C: rgb(255, 0, 0)
D: rgb(0, 0, 0)
PASS successfullyParsed is true

TEST COMPLETE
AAA
BBB
DDD
